feat: "경희대·한국외대 구성원 인증" 서약 페이지 제작
