+++ src/pages/common/univVerificationStep/FirstPage.tsx
@@ -9,6 +9,7 @@ import { pageFinishAtom } from '~/store/funnel';
 import styled from '@emotion/styled';
 import { css } from '@emotion/react';
 import { colors } from '~/styles/colors';
+import Paddler from '~/components/layout/Pad';
 
 const UNIV_SELECTION_LIST = ['KHU', 'HUFS'] as const;
 
@@ -17,10 +18,8 @@ const FirstPage = () => {
   const setIsPageFinished = useSetAtom(pageFinishAtom);
 
   useEffect(() => {
-    if (univValue) {
-      setIsPageFinished(true);
-    }
-  }, [univValue, setUnivValue]);
+    setIsPageFinished(!!univValue);
+  }, [univValue]);
 
   // jotai 전역 변수가 groupApplyAtom처럼 객체라면 아래 코드처럼 useImmerAtom을 사용해보세요!
   // const [nickname, setNickname] = useImmerAtom(groupApplyAtom);
@@ -31,37 +30,39 @@ const FirstPage = () => {
   // };
 
   return (
-    <Col align={'center'} gap={52}>
-      {/*<div>{nickname.info_name.data}</div>*/}
-      {/*<div onClick={handleSetNickname}>useImmerAtom으로 전역 객체 관리해보기 </div>*/}
-      <Text
-        label={'본인 학교를 선택해 주세요 (˵^࿄^˵)'}
-        color={'Gray500'}
-        typography={'NeoTitleM'}
-      />
-      <Row gap={20} justify={'center'}>
-        {UNIV_SELECTION_LIST.map((univ, index) => {
-          return (
-            <UnivSelectContainer
-              key={`${univ} - ${index}`}
-              isClicked={univValue === univ}
-              onClick={() => setUnivValue(univ)}>
-              <IconButton
-                iconName={`univSelection/${univ}`}
-                format={'png'}
-                width={170}
-                height={130}
-              />
-              <Text
-                label={univ === 'KHU' ? '경희대학교' : '한국외국어대학교'}
-                color={'Gray300'}
-                typography={'NeoBodyM'}
-              />
-            </UnivSelectContainer>
-          );
-        })}
-      </Row>
-    </Col>
+    <Paddler top={36} right={20} bottom={24} left={20}>
+      <Col align={'center'} gap={52}>
+        {/*<div>{nickname.info_name.data}</div>*/}
+        {/*<div onClick={handleSetNickname}>useImmerAtom으로 전역 객체 관리해보기 </div>*/}
+        <Text
+          label={'본인 학교를 선택해 주세요 (˵^࿄^˵)'}
+          color={'Gray500'}
+          typography={'NeoTitleM'}
+        />
+        <Row gap={20} justify={'center'}>
+          {UNIV_SELECTION_LIST.map((univ, index) => {
+            return (
+              <UnivSelectContainer
+                key={`${univ} - ${index}`}
+                isClicked={univValue === univ}
+                onClick={() => setUnivValue(univ)}>
+                <IconButton
+                  iconName={`univSelection/${univ}`}
+                  format={'png'}
+                  width={170}
+                  height={130}
+                />
+                <Text
+                  label={univ === 'KHU' ? '경희대학교' : '한국외국어대학교'}
+                  color={'Gray300'}
+                  typography={'NeoBodyM'}
+                />
+              </UnivSelectContainer>
+            );
+          })}
+        </Row>
+      </Col>
+    </Paddler>
   );
 };
