+++ src/pages/common/univVerificationStep/SecondPage.tsx
@@ -1,114 +1,62 @@
-import Text from '~/components/typography/Text';
+import PageLayout from '~/components/layout/page/PageLayout';
+import Paddler from '~/components/layout/Pad';
 import Col from '~/components/layout/Col';
 import Row from '~/components/layout/Row';
-import { css } from '@emotion/react';
-import RoundButton from '~/components/buttons/roundButton/RoundButton';
-import { useInput } from '~/hooks/useInput';
-import { useState } from 'react';
+import { useEffect } from 'react';
 import { useSetAtom } from 'jotai';
+import { personalApplyAtoms } from '~/store/meeting';
 import { pageFinishAtom } from '~/store/funnel';
-import TextInput from '~/components/inputs/textInput/TextInput';
+import Text from '~/components/typography/Text';
+import { css } from '@emotion/react';
 
 const SecondPage = () => {
   const setIsPageFinished = useSetAtom(pageFinishAtom);
-  const { inputValue, _, handleInputChange } = useInput('');
-  const { validateCodeValue, setValidateCodeCleaer, handleValidateCodeValue } =
-    useInput('');
-  const [tryValidate, setTryValidate] = useState(false);
-  const [validateCodeStatus, setValidateCodeStatus] = useState('');
 
-  const handleValidateCodeStatus = (value: string) => {
-    switch (value) {
-      case '':
-        return 'default';
-      case 'error':
-        return `error`;
-      default:
-        return 'default';
-    }
-  };
-  // 인증번호 확인 절차
-  const handleTryValidate = () => {
-    if (inputValue) setTryValidate(true);
-    // 인증코드 api 부착
-  };
-
-  // 인증번호 확인 절차
-  const handleValidate = () => {};
+  useEffect(() => {
+    const isAllInputsFilled = true;
+    setIsPageFinished(!!isAllInputsFilled);
+  }, []);
 
   return (
-    <Col gap={28}>
-      <Col align={'center'} gap={12}>
-        <Text
-          label={'학교 웹메일을 입력해 주세요 (*^࿄^*)'}
-          color={'Gray500'}
-          typography={'NeoTitleM'}
-        />
-        <Text
-          label={
-            '‘시대팅X트로이카’는 시립대, 경희대, 한국외대\n' +
-            ' 학생들에게만 제공되는 서비스입니다.\n' +
-            '해당 학교 구성원임을 인증 후 신청을 진행해 주세요!'
-          }
-          color={'Gray400'}
-          typography={'GoThicBodyS'}
-          css={css`
-            text-align: center;
-          `}
-        />
-      </Col>
-      <Col gap={12}>
-        <Row gap={8}>
-          <TextInput
-            placeholder={'웹메일 주소 입력'}
-            value={inputValue}
-            status={'default'}
-            isAuthentication={true}
-            onChange={handleInputChange}>
+    <Paddler top={36} right={20} bottom={24} left={20}>
+      <Row>
+        <Col gap={12} align="center">
+          <Text
+            label={'웹메일, 전화번호 보안 서약'}
+            color={'Gray500'}
+            typography={'NeoTitleM'}
+            weight={400}
+            size={18}
+          />
+          <Col align="center">
             <Text
-              label={'@khu.ac.kr'}
+              label={
+                '‘시대팅X트로이카’는 시립대, 경희대, 한국외대 학생들에게만 제공되는 서비스입니다.'
+              }
               color={'Gray400'}
-              typography={'GoThicButtonM'}
+              typography={'GoThicBodyS'}
+              weight={400}
+              size={14}
               css={css`
-                padding-left: 5px;
+                text-align: center;
               `}
             />
-          </TextInput>
-          <RoundButton
-            onClick={handleTryValidate}
-            status={inputValue ? 'active' : 'disabled'}
-            borderType={'gray'}
-            height={44}
-            label={''}
-            width={94}>
             <Text
-              label={tryValidate ? '재전송' : '인증하기'}
-              color={'White'}
-              typography={'NeoButtonS'}
+              label={
+                '웹메일 또는 전화번호를 타인에게 양도하거나 도용하는 경우, 향후 서비스 이용에 불이익을 받을 수 있습니다. '
+              }
+              color={'Primary500'}
+              typography={'GoThicBodyS'}
+              weight={400}
+              size={14}
+              css={css`
+                text-align: center;
+              `}
             />
-          </RoundButton>
-        </Row>
-        {tryValidate && (
-          <Row gap={8}>
-            {/*TODO: 인증번호가 숫자인지 문자인지 백엔드쪽에 확인해서 input type 제한 걸기*/}
-            <TextInput
-              placeholder={'인증번호 입력'}
-              value={validateCodeValue}
-              status={handleValidateCodeStatus(validateCodeValue)}
-              isAuthentication={true}
-              onChange={handleValidateCodeValue}
-            />
-            <RoundButton
-              onClick={() => setIsPageFinished(true)}
-              status={'disabled'}
-              height={44}
-              width={94}>
-              <Text label={'확인'} color={'White'} typography={'NeoButtonS'} />
-            </RoundButton>
-          </Row>
-        )}
-      </Col>
-    </Col>
+          </Col>
+        </Col>
+      </Row>
+    </Paddler>
   );
 };
